Enhance cache stats with per-workspace and per-operation breakdowns
- Fix total entry count display (now shows actual cache data vs incorrect counts)
- Add per-workspace statistics with operation breakdowns
- Add global operation statistics across all workspaces
- Remove all legacy cache usage, migrate to universal cache only
- Update daemon to return enhanced CacheStatistics structure
- Maintain backward compatibility with existing cache data
- Add comprehensive statistics collection using list-keys functionality
- Fix clippy collapsible-if warning in database_cache_adapter
- Add dead_code allow attributes for unused legacy methods

+++ lsp-daemon/src/universal_cache/store.rs
@@ -345,7 +345,7 @@ impl CacheStore {
         // === CLEAR FROM DATABASE CACHE ===
 
         // Database: Clear from workspace persistent cache (direct database storage)
-        let mut database_entries_cleared = 0;
+        let mut database_entries_cleared = 0u64;
 
         // Generate all possible cache keys for the symbol
         let mut cache_keys_to_clear = Vec::new();
@@ -509,8 +509,8 @@ impl CacheStore {
         }
 
         // Calculate totals
-        let entries_cleared = database_entries_cleared;
-        let size_freed = (database_entries_cleared * 1024) as u64; // Estimate 1KB per entry
+        let entries_cleared = database_entries_cleared as usize;
+        let size_freed = database_entries_cleared * 1024; // Estimate 1KB per entry
 
         info!(
             "Symbol '{}' cache clearing complete for file {:?}: {} database entries cleared, total={}",
